# MNT Enable O3 optimization for trees