This PR re-enables O3 optimization for trees, because there is a measurable difference between O2 and O3. Here is the result of the ASV benchmark with random forest:

```bash
export SKLBENCH_PROFILE=large_scale
asv continuous  -b RandomForestClassifierBenchmark.time_fit main tree_o3
asv compare main tree_o3
```

```
       before           after         ratio
     [ce00ba81]       [40141f9d]
     <main>       <tree_o3>
       25.5±0.05s       23.5±0.07s     0.92  ensemble.RandomForestClassifierBenchmark.time_fit('dense', 1)
       28.5±0.04s       28.3±0.02s     0.99  ensemble.RandomForestClassifierBenchmark.time_fit('sparse', 1)
```

Note that O3 was enabled by default for trees in `1.1.X`: https://github.com/scikit-learn/scikit-learn/blob/1.1.X/sklearn/tree/setup.py