@glemaitre 2023-01-03
Is it the only estimator that suffers from a regression in the ASV benchmark?