@adrinjalali 2023-01-03
I'm a bit wary of enabling `O3` for anything. An the difference here is not substantial, I would rather stay with `O2`.