@jjerphan 2023-01-03
Thanks you for exploring this again @thomasjpfan.

I think it makes sense to study whether we want to reactivate the third level of optimisation given it was activated in `1.1.X`, as you have mentioned in the description of this PR.

With `gcc`, the optimizers that are being added by upgrading to the third level of optimisation here are:
 - `-fgcse-after-reload`
 - `-fipa-cp-clone`
 - `-floop-interchange`
 - `-floop-unroll-and-jam`
 - `-fpeel-loops`
 - `-fpredictive-commoning`
 - `-fsplit-loops`
 - `-fsplit-paths`
 - `-ftree-loop-distribution`
 - `-ftree-partial-pre`
 - `-funroll-completely-grow-size`
 - `-funswitch-loops`
 - `-fvect-cost-model=dynamic`
 - `-fversion-loops-for-strides`

I think reactivating those is sensible if that it does not enlarge the vendored shared objects too much and if optimizers for other compilers are also free of numerical-approximation.

See this commit for more details about getting this report: https://github.com/scikit-learn/scikit-learn/commit/99e5ebcb5c78416799baf843d0bbf5e386833906