@thomasjpfan 2023-01-03
> Is it the only estimator that suffers from a regression in the ASV benchmark?

From https://scikit-learn.org/scikit-learn-benchmarks/, I do not see anything noticeable.

While refactoring the tree code so that dense + sparse can share code with the best + random splitter, I needed to turn on **O3** to get the same speed as main with **O3**. In summary:

- Tree refactor with **O2** is slower than `main` with **O2**
- Tree refactor with **O3** is faster than `main` with **O2**
- Tree refactor with **O3** is about the same with `main` with **O3**

In above benchmark comparing `main` with **O2** and **O3**, **O3** is about 8% faster than **O2**, I think that is big enough of a performance jump to turn on **O3** for trees.