@jjerphan 2023-01-03
Size of vendored shared objects on `main`:
 - 510 KB `_tree.cpython-311-x86_64-linux-gnu.so`
 - 233 KB `_splitter.cpython-311-x86_64-linux-gnu.so`
 - 193 KB `_utils.cpython-311-x86_64-linux-gnu.so`
 - 258 KB `_criterion.cpython-311-x86_64-linux-gnu.so`

Sizes of vendored shared objects on this branch:
 - 573 KB `_tree.cpython-311-x86_64-linux-gnu.so`
 - 265 KB `_splitter.cpython-311-x86_64-linux-gnu.so`
 - 225 KB `_utils.cpython-311-x86_64-linux-gnu.so`
 - 298 KB `_criterion.cpython-311-x86_64-linux-gnu.so`

So that's an up-to 15% size increase for vendored shared objects.

Do we can similar increase on Mac and Windows? Do we want to afford this cost?
Are there configurations for which we gain massive runtime improvements?