@thomasjpfan 2023-01-03
On MacOS, there is not a big difference: (I do have other processes running since it is my work machine)

```
       before           after         ratio
     [29317601]       [40141f9d]
     <main>           <tree_o3>
       24.7±0.02s       24.5±0.01s     0.99  ensemble.RandomForestClassifierBenchmark.time_fit('dense', 1)
       32.9±0.09s        33.4±0.2s     1.02  ensemble.RandomForestClassifierBenchmark.time_fit('sparse', 1)
```

> Are there configurations for which we gain massive runtime improvements?

On my Linux machine, I have tried other RandomForest configurations and there is always around a 5-8% runtime improvement with **O3**. Overall, I think the 167 KB increase is worth it for the runtime improvement.